+++ pyngrok/ngrok.py
@@ -27,7 +27,7 @@ except ImportError:  # pragma: no cover
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "4.0.3"
+__version__ = "4.1.0"
 
 logger = logging.getLogger(__name__)